+++ hw-006/src/main/java/ru/glavs/hw005/dao/CommentDao.java
@@ -7,5 +7,5 @@ public interface CommentDao {
 
     void delete(Comment comment);
 
-    Comment getById(int id);
+    Comment getById(long id);
 }
